Fix for Makefile GO_TEST_FLAGS, GO_TEST, (etc) overrides, borked by go -C ..
The -C option must be first. I also didn't properly consider that users might
need to know the directory to change to. Most of the other overrides are after
changing the current working directory. For these overrides, a call-able
pattern has been utilized, instead, accepting the module slug as the first
argument.